Updated file to allow swapping between screens
Only swaps to edit screen for now.  The Create patient screen will be about the same as the edit patient, so I may try to find a way to use the same class but just with a different end result when submitting a patient.  Need to update PatientTable to populate it with information from database instead of a generic Patient class I made.