Removed code used for testing & modified backend
The old patient class was removed from the front end since we will be using the backend's object from here on.  The backend was updated so that the database does not recreate itself from scratch every time create_table is called, but checks if the file exists first before creating a new database.